requirements.txt
streamlit>=1.28.0
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.0
pymupdf>=1.23.0
rank-bm25>=0.2.2
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
